added mixin and fixed agent_moved location
